[refactor] Move `GridImageFileStore` to single file
