+++ grid-cell/grid-cell/Cache/GridImageDataStore.swift
@@ -15,23 +15,3 @@ protocol GridImageDataStore {
 enum FileStore {
   static let storeURL = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first!
 }
-
-class GridImageFileStore: GridImageDataStore {
-  let storeURL: URL
-  
-  init(storeURL: URL = FileStore.storeURL) {
-    self.storeURL = storeURL
-  }
-  
-  func insert(_ data: Data, for url: URL) throws {
-    try data.write(to: composeStorePath(with: url))
-  }
-  
-  func retrieve(from url: URL) throws -> Data? {
-    try Data(contentsOf: composeStorePath(with: url))
-  }
-  
-  private func composeStorePath(with url: URL) -> URL {
-    storeURL.appendingPathComponent(url.lastPathComponent)
-  }
-}
